[server] Include server in samlang
In the past, the server code lives in my website repo's `samlang-demo-backend`. The plan is to centralize the samlang Kotlin code into one place. Although this move will increase the binary size, we eventually need this since LSP is on the roadmap.
I also added a script to deploy it to my google cloud run instance.