[checker] Extract a completely immutable `AccessibleGlobalTypingContext`
We want to keep the local typing context thin so we can do some refactoring to make it stack based.
Here is an observation: the global context extracted in this diff is always immutable while we are type checking an expression as a body of a function/method. Let's take it out!