[test] Unbreak master build
This is actually a bug in the IR lowering. The current strategy of lowering cannot rename the variabled in nested scope to prevent this kind of collision.
However, given that I plan to add statements eventually to eradicate this kind of problems, it's not worth putting tons of time to fix it. Thus, I just changed the tests.