[syntax] Nuke `unit` literal and revamp lowering involving unit
In theory, no useful pure function should involve any unit. Yet, we still need to support it in order to be practical. However, with the introduction of statement block, an expression of unit can be implicitly created via `{}`, so the unit literal is unneccessary.
In the past, keeping the `unit` literal has resulted in some clever but actually broken hacks during IR lowering, so its harm actually outweights its benefits. Thus, let's nuke it once and for all to fix the broken code.

Now with `unit` removed, it's the time to cleanup the expression lowering code. If we make the lowered expression in the result nullable, we make many flows explicit and suddenly eliminated a lot of hacks. As a result, Java target compilation involves unit is no longer broken!