[compiler] Always compile the entire source set
We need this to simplify our code so that we can implement a working version of multiple targets for CLI elegantly.
However, this feature imposes serious burden on testing. The reason is that we have now to optimize a large chunk of unrelated merged functions, even if they are unused. To help speed up testing, an optimizer that removed unused functions are implemented.