[checker] Unbreak private class type checking (#37)
### Summary

Private class support seems to be broken since it was introduced. The problem is that the private class is NOT registered in the global typing context. Therefore, when you try to use a private class within a module, it will fail. It's not caught in a test even though we have a test on private classes, because in that test we never tried to use the private classes.

In this diff, we no longer ignore private classes in global typing context. We don't need to worry about accidentally using private classes in another module, because imports of private classes are checked:

https://github.com/SamChou19815/samlang/blob/f7bade8fecd27429154b4d7d15534eef3b610b7c/samlang-core/src/checker/undefined-imports-checker.ts#L30-L33

### Test Plan

Modified the test to include use of private classes. It still passed.