[printer][prettier] Implement `LINE_FLATTEN_TO_NIL` as a prettier extension
### Summary

The semantics of the construst is already documented in the code, so I will not repeat it here. Addition of this construct is necessary to get a nicer looking output in the following scanario:

Suppose we want to pretty print a list. When we have enough space, we can print it like:

```
[a, b, c, d, e, f]
```

When we don't have enough space, we can indent the elements in the list like:

```
[
  a,
  b,
  c,
  d,
  e,
  f
]
```

In the prettier core algorithm, the conditional space or indentation is represented by the `LINE` document. However, the `LINE` document always flattens to a space. Therefore, in the flatten output, we will actually get:

```
[ a, b, c, d, e, f ]
```

, which isn't what we expect.

### Test Plan

See the modified test. Now it no longer prints extra whitespaces.