[printer][source-level] Omit undediced type printing in optional annotations
Due to type inference, type annotation is not always required. In those places where it's not required, we current still print the type during pretty printing. This is bad because now pretty printing can only be done after type checking.

This diff removes this restriction by avoid printing the type annotation when they are optional and the type in the AST is undecided.