[ast][parsing] Increase statement block precedence (#151)
Statement Block doesn't introduce a lot of parsing ambiguity. Giving it a low precedence seems to be a leftover relic of it being part of `val` expressions. It also gives weird pretty print results.

This diff lifts its precedence level to be just behind object constructor. The rest of the changes are just autoformatting.