[ast] Eliminate pointer type from HIR
HIR should be able to be compiled to TS directly. Therefore, a pointer type there doesn't make sense. In addition, we know for sure that only function type and tuple type needs to be wrapped by pointer type, so we can only start to do that at LLVM transformation phase.