[ast][compiler][optimization][hir] Adjust while AST (#338)
## Summary

The current while AST avoids the `break` instruction, but comes with the cost of being very cumbersome. The convoluted way of producing the break condition probably will hinder future potential optimizations. Therefore, this diff aims to fix this by refactoring the AST to make break more explicit.

In this diff, I make them attached to if-else, and switch, but I will explore whether this restriction should be removed.

## Test Plan

- `yarn test`
- `yarn test:integration`