[compiler][mir] Add more dynamic ref counting functions
In this diff, we replace the current static inc and dec ref counting function with a very dynamic ones. With the previous restriction on struct size, we can now encode the struct header as follows:

```
| 16 bits of bitset | 16 bits of ref count |
```

where the first 16 bits encodes whether a field is a ref type. This information can then be used to help us decide at runtime whether to recursively call dec_ref function. This implementation helps to significantly reduce the code size, and helps with future integration with WASM gc as well.