[ast] Introduce HIR and optimize optimizer perf (#1007)
It's the time to reintroduce HIR, as a prep step to perform enum
representation optimization during generics specialization.

The above-mentioned optimization is not done yet. Instead, this commit
focuses on getting the painful refactor done first.

Now that generics specialization transforms HIR into MIR, MIR no longer
needs to be generic. This suddenly makes the MIR expression copyable,
which leads to a serious of perf improvements. We are able to cut 40s of
runtime out of 220s on 10k repeated runs.