[repo] Delay encoding of type and function names until printing stage (#1027)
This is the diff that takes a lot of effort to get rid of the old bad
habit of storing things in strings in unnecessary ways. The performance
aspects are mostly mitigated by the use of PStrs, but it's still very
confusing, since there is no type safety between different kinds of
strings.

In this diff, I create two struct encoded names: TypeName and
FunctionName. TypeName has quite a few fields, so copying is bad for
perf. Therefore, I have to create another lookup table (SymbolTable)
similar to the string heap to mitigate this, so that it can be 4 byte
instead of 48 byte.